# openapi_spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BNC Digital Library - Advanced API",
    "description": "API pour les fonctionnalités avancées du système BNC",
    "version": "1.0.0",
    "contact": {
      "name": "BNC Support",
      "email": "support@bnc-library.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Development Server"
    },
    {
      "url": "https://api.bnc-library.com/api",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/advanced/recommendations/personalized/": {
      "get": {
        "summary": "Obtenir les recommandations personnalisées",
        "description": "Récupère une liste personnalisée de recommandations basée sur le ML",
        "operationId": "get_personalized_recommendations",
        "tags": ["Recommendations"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Nombre de recommandations à retourner",
            "schema": {"type": "integer", "default": 10, "maximum": 100}
          },
          {
            "name": "include_trending",
            "in": "query",
            "description": "Inclure les livres en tendance",
            "schema": {"type": "boolean", "default": true}
          },
          {
            "name": "recommendation_type",
            "in": "query",
            "description": "Type de recommandation",
            "schema": {
              "type": "string",
              "enum": ["collaborative", "content_based", "hybrid", "trending"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des recommandations personnalisées",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {"type": "integer"},
                    "results": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Recommendation"}
                    }
                  }
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Recommendation": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "book": {"$ref": "#/components/schemas/Book"},
          "recommendation_type": {
            "type": "string",
            "enum": ["collaborative", "content_based", "hybrid", "trending"]
          },
          "score": {"type": "number", "minimum": 0, "maximum": 100},
          "is_viewed": {"type": "boolean"},
          "is_liked": {"type": "boolean"},
          "is_purchased": {"type": "boolean"},
          "is_read": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        },
        "required": ["id", "book", "recommendation_type", "score"]
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "title": {"type": "string"},
          "isbn": {"type": "string"},
          "authors": {
            "type": "array",
            "items": {"type": "string"}
          },
          "cover_url": {"type": "string", "format": "uri"},
          "language": {"type": "string"},
          "description": {"type": "string"}
        },
        "required": ["id", "title", "isbn"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "detail": {"type": "string"}
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Non authentifié",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"}
          }
        }
      },
      "Forbidden": {
        "description": "Accès refusé",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"}
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}